time: change unix timestamp to u64 (closes #4258)

+++ vlib/time/time.v
@@ -31,7 +31,7 @@ pub:
 	hour   int
 	minute int
 	second int
-	unix int
+	unix   u64
 }
 
 pub enum FormatTime {
